+++ src/com/kaylerrenslow/armaDialogCreator/data/HeaderConversionException.java
@@ -1,0 +1,11 @@
+package com.kaylerrenslow.armaDialogCreator.data;
+
+/**
+ @author Kayler
+ @see HeaderToProject
+ @since 04/30/2017 */
+public class HeaderConversionException extends Exception {
+	public HeaderConversionException(String message) {
+		super(message);
+	}
+}